# Conflicts:
#	build.gradle
#	gradle.properties
#	src/generated/resources/data/betterarcheology/loot_tables/blocks/evoker_trap.json
#	src/main/java/net/Pandarix/betterarcheology/BetterArcheology.java
#	src/main/java/net/Pandarix/betterarcheology/block/ModBlocks.java
#	src/main/java/net/Pandarix/betterarcheology/block/entity/ArcheologyTableBlockEntity.java
#	src/main/java/net/Pandarix/betterarcheology/compat/jei/IdentifyingCategory.java
#	src/main/java/net/Pandarix/betterarcheology/enchantment/SoaringWindsEnchantment.java
#	src/main/java/net/Pandarix/betterarcheology/util/ArtifactEnchantmentHelper.java
#	src/main/java/net/Pandarix/betterarcheology/villager/ModVillagers.java
#	src/main/resources/META-INF/accesstransformer.cfg
#	src/main/resources/data/betterarcheology/recipes/penetrating_strike_from_identifying.json
#	src/main/resources/data/betterarcheology/recipes/soaring_winds_from_identifying.json
#	src/main/resources/data/betterarcheology/recipes/tunneling_from_identifying.json
#	src/main/resources/data/betterarcheology/structures/archeologist_camp_grassy.nbt
#	src/main/resources/data/betterarcheology/structures/archeologist_camp_redsand.nbt
#	src/main/resources/data/betterarcheology/structures/archeologist_camp_sand.nbt
#	src/main/resources/data/betterarcheology/structures/buried_ruins_sand.nbt
#	src/main/resources/data/betterarcheology/structures/desert_obelisk.nbt
#	src/main/resources/data/betterarcheology/structures/fossil_chicken.nbt
#	src/main/resources/data/betterarcheology/structures/fossil_chicken_birch.nbt
#	src/main/resources/data/betterarcheology/structures/fossil_creeper.nbt
#	src/main/resources/data/betterarcheology/structures/fossil_jungle_0.nbt
#	src/main/resources/data/betterarcheology/structures/fossil_jungle_1.nbt
#	src/main/resources/data/betterarcheology/structures/fossil_sheep_0.nbt
#	src/main/resources/data/betterarcheology/structures/mesa_ruins.nbt
#	src/main/resources/data/betterarcheology/structures/mott.nbt
#	src/main/resources/data/betterarcheology/structures/ruins_sand.nbt
#	src/main/resources/data/betterarcheology/structures/stonehenge_grassy.nbt
#	src/main/resources/data/betterarcheology/structures/temple_jungle.nbt
#	src/main/resources/data/betterarcheology/structures/tumulus_grassy.nbt
#	src/main/resources/data/betterarcheology/structures/underwater_0.nbt
#	src/main/resources/data/betterarcheology/structures/underwater_1.nbt
#	src/main/resources/data/betterarcheology/structures/underwater_2.nbt
#	src/main/resources/data/betterarcheology/structures/underwater_3.nbt
#	src/main/resources/data/betterarcheology/structures/villager_grave.nbt
#	thanks.md

+++ src/main/java/net/Pandarix/betterarcheology/mixin/FasterBrushingMixin.java
@@ -11,13 +11,17 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(BrushableBlockEntity.class)
-public abstract class FasterBrushingMixin {
-    @Shadow private long coolDownEndsAtTick;
+public abstract class FasterBrushingMixin
+{
+    @Shadow
+    private long coolDownEndsAtTick;
 
     @Inject(method = "brush", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/block/entity/BrushableBlockEntity;unpackLootTable(Lnet/minecraft/world/entity/player/Player;)V"))
-    private void injectMethod(long worldTime, Player player, Direction hitDirection, CallbackInfoReturnable<Boolean> cir){
-        if(player.getUseItem().getItem() instanceof BetterBrushItem brushItem){
-            this.coolDownEndsAtTick -= (10L - brushItem.getBrushingSpeed());
+    private void injectMethod(long worldTime, Player player, Direction hitDirection, CallbackInfoReturnable<Boolean> cir)
+    {
+        if (player.getUseItem().getItem() instanceof BetterBrushItem brushItem)
+        {
+            this.coolDownEndsAtTick -= (long) (10L - brushItem.getBrushingSpeed());
         }
     }
 }
